more refactoring; created display.cpp and .h files
