+++ desmume/src/frontend/windows/DeSmuME.vcxproj
@@ -285,6 +285,7 @@
     <ClCompile Include="console.cpp" />
     <ClCompile Include="CWindow.cpp" />
     <ClCompile Include="ddraw.cpp" />
+    <ClCompile Include="display.cpp" />
     <ClCompile Include="FEX_Interface.cpp" />
     <ClCompile Include="File_Extractor\7z_C\7zAlloc.c" />
     <ClCompile Include="File_Extractor\7z_C\7zBuf.c" />
@@ -585,6 +586,7 @@
     <ClInclude Include="console.h" />
     <ClInclude Include="CWindow.h" />
     <ClInclude Include="ddraw.h" />
+    <ClInclude Include="display.h" />
     <ClInclude Include="FEX_Interface.h" />
     <ClInclude Include="File_Extractor\7z_C\7z.h" />
     <ClInclude Include="File_Extractor\7z_C\7zAlloc.h" />
